add command setter with different user role change data types for data_models

+++ core/database/data_models.py
@@ -1,5 +1,5 @@
 from sqlalchemy.orm import relationship
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, func, Table
+from sqlalchemy import Column, BigInteger, SmallInteger, String, ForeignKey, DateTime, func, Table
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -10,16 +10,16 @@ Base = declarative_base()
 class User(Base):
     __tablename__ = 'users'
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     name = Column(String, nullable=False)
-    balance = Column(Integer, default=0)
+    balance = Column(SmallInteger, default=0)
     created_at = Column(DateTime, default=func.now())
 
 
 class Admin(Base):
     __tablename__ = 'admins'
 
-    admin_id = Column(Integer, primary_key=True)
+    admin_id = Column(BigInteger, primary_key=True)
     name = Column(String, nullable=False)
     created_at = Column(DateTime, default=func.now())
 
@@ -27,17 +27,17 @@ class Admin(Base):
 class WorkoutType(Base):
     __tablename__ = 'workout_types'
 
-    type_id = Column(Integer, primary_key=True)
+    type_id = Column(SmallInteger, primary_key=True)
     type_name = Column(String, nullable=False)
 
 
 class Workout(Base):
     __tablename__ = 'workouts'
 
-    workout_id = Column(Integer, primary_key=True)
-    type_id = Column(Integer, ForeignKey('workout_types.type_id'))
+    workout_id = Column(SmallInteger, primary_key=True)
+    type_id = Column(SmallInteger, ForeignKey('workout_types.type_id'))
     date = Column(DateTime, nullable=False)
-    created_by = Column(Integer, ForeignKey('admins.admin_id'))
+    created_by = Column(BigInteger, ForeignKey('admins.admin_id'))
     created_at = Column(DateTime, default=func.now())
     #
     # type = relationship("WorkoutType")
@@ -47,17 +47,17 @@ class Workout(Base):
 class Status(Base):
     __tablename__ = 'statuses'
 
-    status_id = Column(Integer, primary_key=True)
+    status_id = Column(SmallInteger, primary_key=True)
     status_name = Column(String, nullable=False)
 
 
 class Registration(Base):
     __tablename__ = 'registrations'
 
-    registration_id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey('users.user_id'))
-    workout_id = Column(Integer, ForeignKey('workouts.workout_id'))
-    status_id = Column(Integer, ForeignKey('statuses.status_id'), default=1)
+    registration_id = Column(SmallInteger, primary_key=True, autoincrement=True)
+    user_id = Column(BigInteger, ForeignKey('users.user_id'))
+    workout_id = Column(SmallInteger, ForeignKey('workouts.workout_id'))
+    status_id = Column(SmallInteger, ForeignKey('statuses.status_id'), default=1)
     registered_at = Column(DateTime, default=func.now())
 
     # user = relationship("User")
@@ -68,9 +68,9 @@ class Registration(Base):
 class AttendanceHistory(Base):
     __tablename__ = 'attendance_history'
 
-    history_id = Column(Integer, primary_key=True)
-    user_id = Column(Integer, ForeignKey('users.user_id'))
-    workout_id = Column(Integer, ForeignKey('workouts.workout_id'))
+    history_id = Column(SmallInteger, primary_key=True)
+    user_id = Column(BigInteger, ForeignKey('users.user_id'))
+    workout_id = Column(SmallInteger, ForeignKey('workouts.workout_id'))
     attended_at = Column(DateTime, default=func.now())
 
     # user = relationship("User")
